+++ perl-tree-sitter.novaextension/queries/symbols.scm
@@ -1,7 +1,24 @@
+; Technically functions in a package are methods too, but there is no way to disambiguate between them for now
+; (source_file
+;   (package_statement)
+; (subroutine_declaration_statement
+;     name: (bareword) @name
+;    (#set! role method)
+;   ) @subtree
+; )
+
 (subroutine_declaration_statement
   name: (bareword) @name
  (#set! role function)
 ) @subtree
+
+; ambiguous_function_call_expression with a string literal then sub is what Mojolicious routes look like
+(ambiguous_function_call_expression
+  (function) @name
+  (string_literal) @name
+  (anonymous_subroutine_expression
+  (#set! role function)
+)) @subtree
 
 (source_file
   (package_statement